requirements.txt
ace_tools==0.0
asttokens==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
prompt_toolkit==3.0.51
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.1
rdkit==2024.9.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==80.1.0
six==1.17.0
sklearn-compat==0.1.3
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
